Close #122
Close #122 Corrected issue where selecting Chainfist option for a terminator in a terminator squad would caause Chainfist to be selected for every terminator in the squad.

Chainfist was linked to a collective version. Re-linked to a non local shared version. Additionally changed Power fist link to non local shared version.