Close #194
Close #194
Unt weapon choice for immortals was not accounting for the initial cost of a single weapon when incrementing the cost of the weapon choice for every Immortal model in the unit. Solved simply by decremenitng the point cost by 9 initially.

This is most likely not the most elegent  solution and will have to be modified if the point cost of the weapons ever change.