fix(Imperial Fists):
Explusiaris is an Exorcists (Imperial Fists successors) that replace a
Power Sword or a Master-Crafted Power Sword. The Indomitus Captain
(Primaris Captain with Master-crafted Sword and Relic Shield) have a
special wargear Selection with nested wargear, and the relic selector
does seem to work on it.

There is in fact multiple definitions of the same sword.

As a fix, this adds all the swords to the relic conditions

Closes: #10046